requirements.txt
Django==3.1.3
psycopg2==2.8.6
django-compressor==2.4
django-axes==5.8.0
django-two-factor-auth==1.13
phonenumberslite==8.12.12
django-reversion==3.0.8
django-recaptcha==2.0.6
django-csp==3.7
django-csp-reports==1.4
dateparser==1.0.0
cryptography==3.2.1
data-wizard==1.3.0
social-auth-app-django==4.0.0
